separate mml import/export methods
