+++ packages/avatar-creator/src/components/MmlOverlayExport.tsx
@@ -12,16 +12,11 @@ import * as React from "react";
 import { MouseEvent, useEffect, useRef } from "react";
 
 import { AvatarLoader } from "../scripts/avatar-loader";
+import { mmlExport } from "../scripts/mml";
 import { MmlOverlay } from "./MmlButtons";
 import styles from "./MmlOverlayExport.module.css";
 
-
 hljs.registerLanguage("xml", xml);
-
-const keyReplace = {
-  "top:secondary": "topSecondary",
-  "bottom:secondary": "bottomSecondary",
-};
 
 export default function MmlOverlayExport({
   setActive,
@@ -38,24 +33,7 @@ export default function MmlOverlayExport({
 
     hljs.highlightElement(codeRef.current);
 
-    const className = [
-      avatarLoader.getBodyType(),
-      `skin${avatarLoader.getSkin()?.name ?? ""}`,
-    ].join(" ");
-
-    let code = "";
-    code += `<m-character class="${className}" src="${encodeURI(avatarLoader.urls.torso ?? "")}">\n`;
-
-    for (const key in avatarLoader.urls) {
-      if (key === "torso") continue;
-      const url = avatarLoader.urls[key];
-      if (!url) continue;
-      const className = key in keyReplace ? keyReplace[key as keyof typeof keyReplace] : key;
-      code += `    <m-model class="${className}" src="${encodeURI(url)}"></m-model>\n`;
-    }
-
-    code += `</m-character>`;
-
+    let code = mmlExport(avatarLoader);
     code = hljs.highlight(code, { language: "xml" }).value;
 
     codeRef.current.innerHTML = code;
